Services don't need to reload app environment before starting (TORQUE-1049)
Previously we were calling ActiveSupport::Dependencies.clear in Rails
applications before starting each service of an application. This
causes issues if two services are starting at the same time and we
clear everything out while another service is in the middle of
initializating. Because services can't be reloaded without undeploying
and redeploying the application, this was unnecessary and we no longer
call that method when booting a service.

Conflicts:
	modules/services/src/main/java/org/torquebox/services/component/processors/ServiceComponentResolverInstaller.java