Don't accidentally change marshalled string encodings (TORQUE-1073)
When concatenating a literal string with the value produced by
Marshal.dump, it can end up changing the marshalled string encoding
from ASCII-8BIT to UTF-8 or something else, depending on the default
locale. Change our MARSHAL_MARKER to a marshalled string value so that
we aren't changing encodings with concatenation.