[TORQUE-1069] Make it possible to create queues and topic at runtime in the way they are created currently at boot time [TORQUE-1053] Use QueueInstaller to create queues using TorqueBox::Messaging::Queue.start method
This commit fixes both issues. The newly introduced Destinationizer
service is now responsible for creating destinations (queues and topic)
across TorqueBox. It does not talk directly to the HornetQ server and it
utilizes the QueueInstaller (or TopicInstaller) classes from polyglot.

The change itself results in a small overhead when creating the queues
(two MSC services are deployed) but it's a trade-off to have the
destination creation process same across different deployment methods.

This way we end up with a consistent creation of the destinations, both
at deployment and at runtime.

Destinationizer itself executes the creation in an asynchronous way
returning CountDownLatch to optionally wait for the task completion.

Additionally the TorqueBox::Messaging::Queue and
TorqueBox::Messaging::Topic documentation was rewritten to use the
proper yardoc tags.