Expose restart method on shared pools (TORQUE-819)
The restart method, currently exposed via JMX, allows you to send all
new runtime requests for this pool to a new JRuby runtime while
draining old requests off the old runtime.

This commit is enough to get zero-downtime deploy for new application
code in web, scheduled jobs, and messaging runtimes when they're using
a shared pool, which is the default in production mode. It probably
also works for stomplets but that has not been tested yet.

Similar work needs to be done for bounded pools and then something
wired up to look for tmp/restart.txt so this can be triggered outside
of JMX operations.