Fix support for servlets in WEB-INF/web.xml (TORQUE-1087)
Fix support for adding custom Java servlets to a Ruby application by
adding a WEB-INF/web.xml with the appropriate <servlet> and
<servlet-mapping> elements. There's now an integration test verifying
the servlets get created as expected.

Fixing this uncovered another issue where custom servlets that
returned a HTTP response code >= 400 had the response thrown away and
sent to the application's Rack stack instead. This is now fixed as
well.