Fix bug preventing TorqueBox boot on Java 8 (TORQUE-1210)
This upgrades us to a newer AS 7.2 slim build that fixes the
underlying AS7 bug that prevents booting on Java 8. As a side-effect
of that change, we have to remove all usage of
RootedDeploymentProcessor and instead spread that application root
check logic throughout the processors.

This change may also fix our issues with overlaying on EAP 6.2, but
that hasn't been tested.