Fix for using an external jruby with a jar.
Since java defaults to loading classes referenced by class X using the
ClassLoader that loaded X, we have to prevent the base ClassLoader from
seeing classes that depend on java classes if we want to bring in
jruby.jar at runtime. This commit causes wunderboss-rack.jar and
wunderboss-ruby.jar to not be shaded, they are instead just added to the
jar as entries, and loaded via the classpath mechanism provided by
app.properties and ApplicationRunner.

This also treats a bundled jruby.jar the same way so that loading it is
consistent when bundled or external.