Ensure listen always returns a nil value [fixes #204]
the ruby MessageHandler class extends the wboss MessagingHandler class,
which has a return type of Reply. Whatever on_message returns will be
coerced to Reply, which will cause an error & stack in the log if that
fails. Since we don't care about the return value of listen anyway, we
always return nil.