Initial stab at JRuby 9 support for TorqueBox 3.
All integration tests pass as long as first apply
https://github.com/jruby/jruby/pull/3222 and build JRuby from
source. Once that PR is accepted upstream everything should be good
without needing a custom build of JRuby.

This branch of TorqueBox 3 is usable with JRuby 9.0.0.0 out of the
box, but you will run into memory leaks over time as an application is
redeployed until the issue referenced above is fixed.