Ensure repeated war/jar builds don't endlessly grow in size
When excludes were updated to not require a leading slash, the regular
expression we were using to exclude top-level war and jar files from
being bundled inside jar archives broke. Thus, every time a user
bundled a new war or jar we ended up sticking the existing copy of
that war or jar inside the new one. I've added a test to ensure we
don't accidentally break this in the future.

This fixes issues #244.